app/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.1
cycler==0.10.0
Django==3.0.6
django-bootstrap4==1.1.1
django-tempus-dominus==5.1.2.13
gunicorn==20.0.4
kiwisolver==1.2.0
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
pandas==1.0.3
Pillow==7.1.2
plotly==4.8.0
psycopg2==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.1.0
django-ckeditor==6.0.0
psycopg2==2.8.6
whitenoise==5.2.0